fix channels

+++ modules/nf-core/gatk4/markduplicates/main.nf
@@ -44,8 +44,7 @@ process GATK4_MARKDUPLICATES {
         ${reference} \\
         $args
 
-
-    if  [[ ${prefix} == *.cram ]]; then
+    if  [[ ${prefix} == *.cram ]]&&[[ -f ${prefix}.bai ]]; then
         mv ${prefix}.bai ${prefix}.crai
     fi
